+++ app/globals.css
@@ -85,10 +85,14 @@
 }
 
 @font-face {
-  font-family: 'TruenoBlackOutline';
-  src: url('../public/fonts/TruenoBlackOutline.woff2') format('woff2');
+  font-family: "TruenoBlackOutline";
+  src: url("../public/fonts/TruenoBlackOutline.woff2") format("woff2");
 }
 
 .page-heading {
-  font-family: 'TruenoBlackOutline', sans-serif;
+  font-family: "TruenoBlackOutline", sans-serif;
+}
+
+.box-shadow-link {
+  box-shadow: 12px 12px 0px -3px rgba(0, 0, 0, 0.1);
 }
