+++ components/Cards/Talk.tsx
@@ -7,10 +7,14 @@ const PreviousTalk = ({ data }: { data: Talk[] }) => {
   const createCardsList = () => {
     return data.map((item: Talk, index: number) => {
       return (
-        <Link href="#" key={`previousTalk.title + ${index}`}>
-          <div className="carousel-item rounded-lg border-2 border-border p-5 lg:p-7">
-            <div className="md:w-70 md:h-70  flex h-60 w-60 flex-col  xl:h-80 xl:w-80">
-              <div className="flex h-2/3 items-center justify-center">
+        <Link
+          className="min-w-[262px]"
+          href="#"
+          key={`previousTalk.title + ${index}`}
+        >
+          <div className="carousel-item rounded-lg border-2 border-border p-5 transition-all duration-300 hover:shadow-[12px_12px_0px_-3px_#CCCCCC] mb-4 mr-4 lg:p-7">
+            <div className="flex flex-col">
+              <div className="mb-4 flex h-2/3 items-center justify-center">
                 <Image
                   src={item.imgSrc}
                   alt={item.title}
@@ -18,8 +22,12 @@ const PreviousTalk = ({ data }: { data: Talk[] }) => {
                   height={200}
                 />
               </div>
-              <span className="text-[20px] font-medium">{item.date}</span>
-              <h4 className="text-[28px] font-semibold">{item.title}</h4>
+              <span className="text-base font-medium lg:text-[20px]">
+                {item.date}
+              </span>
+              <h4 className="text-[20px] font-semibold lg:text-[28px]">
+                {item.title}
+              </h4>
               <p className="text-text-sub-heading text-lg font-semibold">
                 by {item.speaker}
               </p>
